backend/auth/requirements.txt
psycopg[binary]==3.1.8
bcrypt==4.0.1
PyJWT==2.8.0